+++ CHANGELOG.md
@@ -12,12 +12,14 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 * component detection: use `estraverse` to improve component detection ([#2992][] @Wesitos)
 * [`destructuring-assignment`], [`no-multi-comp`], [`no-unstable-nested-components`], component detection: improve component detection ([#3001][] @vedadeepta)
 * [`no-deprecated`]: fix crash on rest elements ([#3016][] @ljharb)
+* [`destructuring-assignment`]: get the contextName correctly ([#3025][] @ohhoney1)
 
 ### Changed
 * [Docs] [`jsx-no-bind`]: updates discussion of refs ([#2998][] @dimitropoulos)
 * [Refactor] `utils/Components`: correct spelling and delete unused code ([#3026][] @ohhoney1)
 
 [#3026]: https://github.com/yannickcr/eslint-plugin-react/pull/3026
+[#3025]: https://github.com/yannickcr/eslint-plugin-react/pull/3025
 [#3016]: https://github.com/yannickcr/eslint-plugin-react/issues/3016
 [#3006]: https://github.com/yannickcr/eslint-plugin-react/pull/3006
 [#3001]: https://github.com/yannickcr/eslint-plugin-react/pull/3001